src/requirements/requirements-server.txt
playsound==1.2.2
PyYAML
cryptography
pyuac